+++ build-package.ts
@@ -197,7 +197,7 @@ const buildProject = async (): Promise<void> => {
 };
 
 const addTypescriptDefinitions = async () => {
-	const { childProcess } = await nvexeca('16', 'pnpm', [
+	const { childProcess } = await nvexeca('18', 'pnpm', [
 		'tsc',
 		'--project',
 		'./tsconfig.types.json',